Add token verification when requested from mosquitto broker

+++ resources/adapter-scripts/temperature_oauth/install.sh
@@ -2,7 +2,17 @@
 sudo apt-get update;
 sudo apt-get install -y python3;
 sudo apt-get install -y python3-pip;
-pip3 install paho-mqtt;
-pip3 install oauth2;
-pip3 install requests-oauthlib;
-pip3 install requests;
+
+
+#!/bin/bash
+paho_exist=$(pip3 list | grep -F paho-mqtt)
+if [ -z $paho_exist ]; then
+ sudo apt-get update;
+ sudo apt-get install -y python3;
+ sudo apt-get install -y python3-pip;
+ pip3 install paho-mqtt;
+ pip3 install oauth2;
+ pip3 install requests-oauthlib;
+ pip3 install requests;
+fi
+echo "$1 = $2" > $3/connections.txt;
